.claude/skills/api-aggregation-notifier/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
responses>=0.25.0
pytest>=7.0.0
